# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinic/members/invitations:
    post:
      operationId: inviteClinicMember
      tags: [clinic]
      summary: Invite a clinic administrator or physiotherapist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicMemberInviteInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicMemberInvitation"
  /clinic/members/invitations/{token}/accept:
    post:
      operationId: acceptClinicMemberInvitation
      tags: [clinic]
      summary: Accept a clinic invitation
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Membership activated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /clinic/members/{id}:
    patch:
      operationId: updateClinicMember
      tags: [clinic]
      summary: Change a member role or revoke access
      parameters:
        - $ref: "#/components/parameters/RecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicMemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicMember"
  /clinic/assignments:
    get:
      operationId: listClinicAssignments
      tags: [clinic]
      summary: List patient responsibilities
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClinicAssignment"
    post:
      operationId: assignClinicPatient
      tags: [clinic]
      summary: Assign or unassign a physiotherapist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicAssignmentInput"
      responses:
        "200":
          description: Assignment saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicAssignment"
...
  /clinic/appointments/{id}/transfer:
    post:
      operationId: transferClinicAppointment
      tags: [clinic]
      summary: Transfer an appointment to an active physiotherapist
      parameters:
        - $ref: "#/components/parameters/RecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [physiotherapistUserId]
              properties:
                physiotherapistUserId: { type: string }
      responses:
        "200":
          description: Appointment transferred
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  parameters:
...
    RecordId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ClinicMember:
      type: object
      additionalProperties: true
    ClinicMemberInviteInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        role: { type: string, enum: [admin, physiotherapist] }
    ClinicMemberInvitation:
      type: object
      additionalProperties: true
    ClinicMemberUpdate:
      type: object
      properties:
        role: { type: string, enum: [admin, physiotherapist] }
        status: { type: string, enum: [active, revoked] }
    ClinicAssignment:
      type: object
      additionalProperties: true
    ClinicAssignmentInput:
      type: object
      required: [clientId]
      properties:
        clientId: { type: string }
        physiotherapistUserId: { type: ["string", "null"] }
...