# lib/api-zod/src/generated/types/clinicMemberInviteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClinicMemberInviteInputRole } from './clinicMemberInviteInputRole';

export interface ClinicMemberInviteInput {
  email: string;
  role: ClinicMemberInviteInputRole;
}